Backward stable pseudo-inverse for higher accuracy
The matrix for computing check-potential from equivalent density if
ill-conditioned, therefore computing the inverse of the matrix incurs
large errors in finite precision computation. For over a decade, this
seemed to be a limitation of the kernel independent FMM, that we could
not get greater than about 9-digits of accuracy in double precision. The
idea that solved this problem is called a backward-stable pseudo
inverse. This came up in discussion between Alex Barnett and Abtin
Rahimian and was communicated to us by Abtin.

The idea behind backward-stable pseudo-inverse is as follows. We compute
the inverse of a matrix A by first computing its SVD, A=USV'. Then,
compute the inverse A^-1 = V*(S^-1)*U', with proper regularization for
S. The singular values (diagonal elements in S) range over several
orders of magnitude and therefore, when we compute the above product, we
incur errors in finite precision arithmetic (since we add very large and
very small elements). We use regularization to ignore very small values
in S, but this regularization is arbitrary. If our ultimate goal was to
compute the product A^-1*B (or A^-1*v), for some matrix B, which has a
spectrum similar to A (for some vector v which is a linear combination
of the columns of B), then we could us B to regularize. This can be
achieved by computing the product with A^-1 in steps by computing the
product with its factors. The product (S^-1)*U'*B is naturally
regularized by the B (which has a spectrum similar to B. Notice that we
don't need all the three factors, we can either pre-compute V*(S^-1) or
(S^-1)*U'. Notice that this does not work for any arbitrary
ill-conditioned matrices A and B. They need to be related, for example,
they could both be diffusion operators.

The above discussion implies that we can compute operations from
potential-to-potential and from density-to-density with a single
mat-vec. We can also compute the operations from density-to-potential
with a single mat-vec, since it is the forward operation through kernel
evaluation. However, we can not compute the density from potential in
one step. The operation must be represented by its two factors and
therefore requires two mat-vecs. To account for this, we make the
following changes to the FMM.

- For volume FMM, the upward pass remains the same, except that the
  pre-computation of the matrix is done correctly, i.e. in two steps.
  For particle FMM, in source-to-multipole, computing the equivalent
  density from the check potential has to be computed with two mat-vecs.

- For the V-list interactions, computing the equivalent density would
  have required two mat-vecs for each tree-node. Instead, we have
  changed the representation of the local expansion to be the check
  potential (instead of the equivalent density). And the local-to-local
  operation is in terms of check potentials only.

- The local to target step now computes the target potential from the
  potential on the downward check surface. For volume FMM, we require
  only one mat-vec but the pre-computed matrix changes and must be
  computed as discussed above in two steps. For particle FMM, we need to
  compute the equivalent density using two mat-vecs and then compute the
  target potential from that.

- Note that all the pre-computed matrices for volume FMM, will need to
  be re-computed after these changes.

In addition to the above changes, we also make a other minor changes,
discussed below

- examples/include/utils.txx: point_distrib(..) produced the same
  distribution each time since it uses mpi-rank as seed. It now uses a
  static variable to update the seed for subsequent calls.

- Fix un-used variable warning and other warnings, errors with GCC. Also
  requires additional configure flags for GCC (see INSTALL).

- Change notation: Homogeneous kernels are now correctly called
  Scale-Invariant kernels.

- Replace interaction matrix type UC2UE_Type with UC2UE[1,2]_Type for
  backward-stable pseudo-inverse.

+++ src/mem_mgr.cpp
@@ -116,7 +116,7 @@ void MemoryManager::test(){
       tmp=(double*)memgr.malloc(M*sizeof(double)); assert(tmp);
       tt=omp_get_wtime();
       #pragma omp parallel for
-      for(size_t i=0;i<M;i+=64) tmp[i]=i;
+      for(size_t i=0;i<M;i+=64) tmp[i]=(double)i;
       tt=omp_get_wtime()-tt;
       std::cout<<tt<<' ';
       memgr.free(tmp);
@@ -132,7 +132,7 @@ void MemoryManager::test(){
       tmp=(double*)DeviceWrapper::host_malloc(M*sizeof(double)); assert(tmp);
       tt=omp_get_wtime();
       #pragma omp parallel for
-      for(size_t i=0;i<M;i+=64) tmp[i]=i;
+      for(size_t i=0;i<M;i+=64) tmp[i]=(double)i;
       tt=omp_get_wtime()-tt;
       std::cout<<tt<<' ';
       DeviceWrapper::host_free(tmp);
