+++ src/tree_node.cpp
@@ -36,7 +36,7 @@ void TreeNode::Initialize(TreeNode* parent_, int path2node_, NodeData* data_){
     max_depth=parent->max_depth;
   }
 
-  assert(path2node_>=0 && path2node_<(1U<<dim));
+  assert(path2node_>=0 && path2node_<(int)(1U<<dim));
   path2node=path2node_;
 
   //assert(parent_==NULL?true:parent_->Child(path2node_)==this);
